requirements.txt
flask
torch
torchvision
numpy
pillow
matplotlib
gunicorn
